pyproject.toml
[project]
name = "mcp-server-note"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.13.1",
    "supabase>=2.18.1",
    "openai>=1.0.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
